requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20250909
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
asttokens==3.0.1
attrs==25.3.0
blake3==1.0.5
blinker==1.9.0
cachetools==5.5.2
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorful==0.5.7
colorlog==6.9.0
compressed-tensors==0.9.0
cuda-bindings==12.9.6
cuda-core==0.7.0
cuda-pathfinder==1.5.2
cuda-python==12.9.0
cupy-cuda12x==13.6.0
cython==3.2.4
decorator==5.2.1
decord==0.6.0
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.116.2
fastapi-cli==0.0.12
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.19.1
flash==1.0.3
flashinfer-python==0.3.1
flask==3.1.2
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
gguf==0.10.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.75.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
importlib-metadata==8.7.0
interegular==0.3.3
ipython==9.12.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mistral-common==1.8.5
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cccl-cu12==12.9.27
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-frontend==1.14.1
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvfatbin-cu12==12.9.82
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.6.5
nvidia-nvtx-cu12==12.4.127
nvshmem4py-cu12==0.3.0
openai==1.108.0
openai-harmony==0.0.4
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-exporter-prometheus==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
orjson==3.11.3
outlines==0.1.11
outlines-core==0.1.26
packaging==25.0
parso==0.8.6
partial-json-parser==0.2.1.1.post6
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pygments==2.19.2
pynvml==13.0.1
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pyyaml==6.0.2
pyzmq==27.1.0
ray==2.49.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scipy==1.16.2
sentencepiece==0.2.1
sentry-sdk==2.38.0
setproctitle==1.3.7
setuptools==79.0.1
sgl-kernel==0.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.3.1
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
starlette==0.48.0
sympy==1.13.1
tabulate==0.9.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
triton==3.1.0
typer==0.17.4
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
vllm==0.7.1
watchfiles==1.1.0
wcwidth==0.6.0
websockets==15.0.1
werkzeug==3.1.3
wrapt==1.17.3
xformers==0.0.28.post3
xgrammar==0.1.23
yarl==1.20.1
zipp==3.23.0
